unrbeak IPv6 support: change ttl value size back to 32 bit for ipv6
git-svn-id: file:///home/lennart/svn/public/avahi/trunk@853 941a03a8-eaeb-0310-b9a0-b1bbd8fe43fe

+++ avahi-core/socket.c
@@ -307,7 +307,7 @@ fail:
 int avahi_open_socket_ipv6(int no_reuse) {
     struct sockaddr_in6 sa, local;
     int fd = -1, yes, r;
-    uint8_t ttl;
+    int ttl;
 
     mdns_mcast_group_ipv6(&sa);
